bot: major refactoring towards new connection objects The connections refuse to work and I am too lazy and do not intend on loosing my will to live anytime soon, so that is that. Probably contains some other uncaught exceptions as well, I wouldn't know. As a word of caution, do not attempt to debug this code if you are suffering from depression, as that might make it worse. Trust me.

+++ db/setup.sql
@@ -12,7 +12,7 @@ CREATE UNIQUE INDEX IF NOT EXISTS votes_idx ON votes(
 CREATE TABLE IF NOT EXISTS submitted_votes(
 	voter_id INTEGER,
 	server_id INTEGER,
-	vote_timestamp INTEGER
+	vote_timestamp NUMERIC
 );
 
 CREATE UNIQUE INDEX IF NOT EXISTS submitted_votes_idx ON submitted_votes(
